* magick/image.c (SetImageChannelDepth): Depth parameter was being returned rather than status.  Oops! * magick/effect.c (BlurScanline): Due to automatic casting conventions, computation was being done (at least with SGI compiler) as type 'unsigned long' rather than 'long' as it should have been. * coders/jpeg.c, coders/meta.c, coders/miff.c, coders/msl.c, coders/palm.c, coders/pcd.c, coders/psd.c, coders/svg.c, coders/tiff.c, coders/xcf.c, magick/render.c, : Quench many SGI compiler warnings regarding variables which are initialized but never used. * magick/xwindow.h: Undef gravity defines so that enumerated type is used instead.

+++ magick/effect.c
@@ -355,7 +355,7 @@ static void BlurScanline(const double *kernel,const unsigned long width,
               aggregate.blue+=(*p)*q->blue;
               aggregate.opacity+=(*p)*q->opacity;
             }
-          if (((i+width/2-x) >= 0) && ((i+width/2-x) < width))
+          if (((i+(long)(width/2)-x) >= 0) && ((i+width/2-x) < width))
             scale+=kernel[i+width/2-x];
           p++;
           q++;
